Tests/LibWeb/Text/expected/HTML/window-open-about-blank-inherits-opener-origin.txt

initial popup.location.href: about:blank
initial popup.document.body.textContent: Hello Shannon!
reloaded popup.document.body.textContent: 
